source: remove build warning for 64bit build
I fix some pointer casting and printf format warnings for 64bit support.

There are many problems about code convention like indentations,
whitespaces, tabs, and so on. These should be aligned later, not this
commit.

+++ lib/libsms/sms_tool.c
@@ -563,7 +563,7 @@ int MsgConvertUTF8ToGSM7bit(unsigned char *pDestText, int maxLength,  const unsi
     bzero(pUCS2Text, maxUCS2Length);
 
     fprintf(stderr, "srcTextLen = %d\n", srcTextLen);
-    fprintf(stderr, "temp buffer size = %d\n", maxUCS2Length * sizeof(unsigned short));
+    fprintf(stderr, "temp buffer size = %zu\n", (size_t)(maxUCS2Length * sizeof(unsigned short)));
     fprintf(stderr, "max dest Length = %d\n", maxLength);
 
     ucs2Length = MsgConvertUTF8toUCS2((unsigned char*)pUCS2Text, maxUCS2Length * sizeof(unsigned short), pSrcText, srcTextLen);
